made several changes
reformated the file strucure of the controllers

fixed the bug where every time the server want to send message to the RMQ it establishes new connection.
