+++ src/services/User Management/rmq/pushToRMQ.js
@@ -1,10 +1,7 @@
-const amqp = require("amqplib");
-
+const { getChannel } = require("../rmq/establishRMQConnection");
 const pushToQueue = async (state, followerId, followingId) => {
+  const channel = getChannel();
   try {
-    const connection = await amqp.connect(process.env.RMQURL);
-    const channel = await connection.createChannel();
-    await channel.assertQueue("following-management", { durable: true });
     const message = {
       state: state,
       followerId: followerId,
@@ -15,8 +12,6 @@ const pushToQueue = async (state, followerId, followingId) => {
       Buffer.from(JSON.stringify(message))
     );
     console.log("Message pushed to following-management queue", message);
-    await channel.close();
-    await connection.close();
   } catch (err) {
     throw new Error(
       `Error happened while trying to push the message to the following-management queue: ${err}`
